Create distributed_processing.h
This code uses MPI to distribute the workload among multiple processes. The master process (rank 0) distributes the workload among the worker processes by iterating through the total amount of work and assigning each iteration to the next process. Each process performs the assigned task and sends the result back to the master process. The master process collects the results from the worker processes and process it.

This is a simple example that demonstrates the basic idea of how to use MPI to distribute the workload among multiple processes. In a real-world application, you may need to adjust this code to handle more complex workloads, such as distributing large data sets or implementing more advanced communication patterns.

It's important to note that in order to run this code you need to have MPI library installed in your system, also you need to use an MPI compiler to build and run the program, such as mpic++.